Fix Filter package structure
